refactor(InputCaption): moved to internal
