+++ src/internal/components/InputCaption.tsx
@@ -1,6 +1,6 @@
 import React from 'react'
-import Text from './Text'
-import {SxProp} from './sx'
+import Text from '../../Text'
+import {SxProp} from '../../sx'
 
 type Props = {
   /**
